Allow for running 32-bit apps in 32-bit processes
Use this by default instead of 32in64

+++ src/dyld/gdb.h
@@ -26,7 +26,7 @@ along with Darling.  If not, see <http://www.gnu.org/licenses/>.
 extern "C" {
 #endif
 
-void setup_gdb_notifications(uint64_t slide, uint64_t addr);
+void setup_gdb_notifications(uintptr_t slide, uintptr_t addr);
 
 #ifdef __cplusplus
 }
